+++ CHANGES.md
@@ -2,6 +2,7 @@
 
 ## Version 2.11.0
 
+* Added search option to user view
 * Added support for discarding sessions. Admins can now mark sessions for deletion after a certain period of time (30 days by default). This is useful for temporary sessions that are not interesting for keeping forever.
 * Sessions can have a TTL, automatically marking their discard date on keepalives and session ends
 * Support reporting of remote and local SCM branches